Fix compiler warnings.

+++ test/Random/RandomGeneratorTest.cpp
@@ -34,7 +34,7 @@ TEST_F( RandomGeneratorTest, getTest ) {
 
     static const size_t Size = 100;
     int old( -10 );
-    for ( int i = 0; i < Size; i++ ) {
+    for ( size_t i = 0; i < Size; i++ ) {
         int v = generator.get( 0, 10 );
         if ( -10 != old ) {
             EXPECT_NE( old, v );
